refactor(backup): unify --clobber backup into one shared helper
`wt switch --clobber` and `wt step relocate --clobber` each had their own
backup-the-blocker logic. They diverged: switch's `back_up_clobbered_path`
iterates suffixes (`…-2`, `…-3`, …) on a name collision so it never fails,
while relocate's `backup_blocking_path` failed closed on `AlreadyExists`.
A relocate run could abort on a same-second clobber that switch handles
gracefully.

Both now go through `commands::backup` — `back_up_clobbered_path_now`
computes the timestamp suffix and delegates to the suffix-iterating
`back_up_clobbered_path`. relocate gains the suffix fallback, the
timestamp computation is no longer duplicated across the two call sites,
and `renamore::rename_exclusive` has a single caller.

relocate's backup name changes from `.bak-<timestamp>` to the
extension-aware `.bak.<timestamp>` form (e.g. `repo.feature` →
`repo.feature.bak.<timestamp>`), matching switch. The
`test_relocate_clobber_error_backup_exists` regression test asserted the
old fail-closed behavior; it is rewritten as
`test_relocate_clobber_falls_back_when_backup_taken` to assert the
corrected `-N` fallback, mirroring the switch test. CLI help and the
generated doc mirrors are updated for the new naming and fallback.
