+++ src/commands/mod.rs
@@ -1,4 +1,5 @@
 mod alias;
+pub(crate) mod backup;
 pub(crate) mod command_approval;
 pub(crate) mod command_executor;
 pub(crate) mod commit;
